Add non-amd webpack output
We need to support scenarios when consumer cannot use the nuget package
which contains the amd module.
In such scenarios they will have to install an npm module and
initialize on their own.

This commit introduces typescript together with definition generation so
that it's easier to use by consumers.
AMD version of tracker factory is just a proxy to the base factory
exposed as a singleton so that the current consumers (e.g. Labs) would
not have to change their code.

Task: CMS-15849
Feature: CMS-15839