Allow to create factory when no config available yet
It should be possible to create a factory instance without the config.
Since the initialize is async we need explicit initialize method on
the factory. The factory will behave as Null-Value factory until it's
initialized.
The AMD module will just proxy to that new method.
The constructor accepts a configuration parameter in case the consumer
already has the config while creating the factory instance.

Task: CMS-15876